Clean up and fix handling of environment
