+++ src/vs/workbench/parts/terminal/node/terminalProcess.ts
@@ -7,14 +7,10 @@ import * as os from 'os';
 import * as path from 'path';
 import * as platform from 'vs/base/common/platform';
 import * as pty from 'node-pty';
-import * as terminalEnvironment from 'vs/workbench/parts/terminal/node/terminalEnvironment';
 import { Event, Emitter } from 'vs/base/common/event';
 import { ITerminalChildProcess } from 'vs/workbench/parts/terminal/node/terminal';
 import { IDisposable } from 'vs/base/common/lifecycle';
-import { IShellLaunchConfig, ITerminalConfigHelper } from 'vs/workbench/parts/terminal/common/terminal';
-import { IWorkspaceContextService, IWorkspaceFolder } from 'vs/platform/workspace/common/workspace';
-import { IConfigurationResolverService } from 'vs/workbench/services/configurationResolver/common/configurationResolver';
-import URI from 'vs/base/common/uri';
+import { IShellLaunchConfig } from 'vs/workbench/parts/terminal/common/terminal';
 
 export class TerminalProcess implements ITerminalChildProcess, IDisposable {
 	private _exitCode: number;
@@ -32,14 +28,11 @@ export class TerminalProcess implements ITerminalChildProcess, IDisposable {
 	public get onProcessTitleChanged(): Event<string> { return this._onProcessTitleChanged.event; }
 
 	constructor(
-		@IWorkspaceContextService private readonly _workspaceContextService: IWorkspaceContextService,
-		@IConfigurationResolverService private readonly _configurationResolverService: IConfigurationResolverService,
 		shellLaunchConfig: IShellLaunchConfig,
 		cwd: string,
 		cols: number,
 		rows: number,
-		lastActiveWorkspaceRootUri: URI,
-		private readonly _configHelper: ITerminalConfigHelper
+		env: platform.IProcessEnvironment
 	) {
 		// The pty process needs to be run in its own child process to get around maxing out CPU on Mac,
 		// see https://github.com/electron/electron/issues/38
@@ -52,13 +45,6 @@ export class TerminalProcess implements ITerminalChildProcess, IDisposable {
 			shellName = 'xterm-256color';
 		}
 
-		if (!shellLaunchConfig.executable) {
-			this._configHelper.mergeDefaultShellPathAndArgs(shellLaunchConfig);
-		}
-
-		const lastActiveWorkspaceRoot = this._workspaceContextService.getWorkspaceFolder(lastActiveWorkspaceRootUri);
-
-		const env = this._createEnv(shellLaunchConfig, lastActiveWorkspaceRoot);
 		const options: pty.IPtyForkOptions = {
 			name: shellName,
 			cwd,
@@ -92,58 +78,6 @@ export class TerminalProcess implements ITerminalChildProcess, IDisposable {
 		this._onProcessExit.dispose();
 		this._onProcessIdReady.dispose();
 		this._onProcessTitleChanged.dispose();
-	}
-
-	private _createEnv(shellLaunchConfig: IShellLaunchConfig, lastActiveWorkspaceRoot: IWorkspaceFolder): platform.IProcessEnvironment {
-		// TODO: Move locale into TerminalProcess
-		const locale = this._configHelper.config.setLocaleVariables ? platform.locale : undefined;
-		// Resolve env vars from config and shell
-		const platformKey = platform.isWindows ? 'windows' : (platform.isMacintosh ? 'osx' : 'linux');
-		const envFromConfig = terminalEnvironment.resolveConfigurationVariables(this._configurationResolverService, { ...this._configHelper.config.env[platformKey] }, lastActiveWorkspaceRoot);
-		const envFromShell = terminalEnvironment.resolveConfigurationVariables(this._configurationResolverService, { ...shellLaunchConfig.env }, lastActiveWorkspaceRoot);
-		shellLaunchConfig.env = envFromShell;
-
-		const env: platform.IProcessEnvironment = { ...process.env };
-
-		// Merge process env with the env from config
-		// TODO: Move environment merge stuff into TerminalProcess
-		terminalEnvironment.mergeEnvironments(env, envFromConfig);
-
-		// Continue env initialization, merging in the env from the launch
-		// config and adding keys that are needed to create the process
-		const env = terminalEnvironment.createTerminalEnv(shellLaunchConfig, locale);
-
-		const keysToRemove = [
-			'ELECTRON_ENABLE_STACK_DUMPING',
-			'ELECTRON_ENABLE_LOGGING',
-			'ELECTRON_NO_ASAR',
-			'ELECTRON_RUN_AS_NODE',
-			'GOOGLE_API_KEY',
-			'VSCODE_CLI',
-			'VSCODE_DEV',
-			'VSCODE_IPC_HOOK',
-			'VSCODE_LOGS',
-			'VSCODE_NLS_CONFIG',
-			'VSCODE_PORTABLE',
-			'VSCODE_PID',
-		];
-		keysToRemove.forEach((key) => {
-			if (env[key]) {
-				delete env[key];
-			}
-		});
-		Object.keys(env).forEach(key => {
-			if (key.search(/^VSCODE_NODE_CACHED_DATA_DIR_\d+$/) === 0) {
-				delete env[key];
-			}
-		});
-		// TODO: Determine which parts of env initialization should go where
-		// const envFromConfig = terminalEnvironment.resolveConfigurationVariables(this._configurationResolverService, { ...this._configHelper.config.env[platformKey] }, lastActiveWorkspaceRoot);
-		// const envFromShell = terminalEnvironment.resolveConfigurationVariables(this._configurationResolverService, { ...shellLaunchConfig.env }, lastActiveWorkspaceRoot);
-		// shellLaunchConfig.env = envFromShell;
-		// terminalEnvironment.mergeEnvironments(parentEnv, envFromConfig);
-		// const env = terminalEnvironment.createTerminalEnv(parentEnv, shellLaunchConfig, this.initialCwd, locale, cols, rows);
-		return env;
 	}
 
 	private _setupTitlePolling() {
